requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==4.2.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
cssselect==1.1.0
dateparser==1.1.1
demjson3==3.0.5
facebook-scraper==0.2.56
fake-useragent==0.1.11
idna==3.3
importlib-metadata==4.11.4
lxml==4.9.0
numpy==1.22.4
pandas==1.4.2
parse==1.19.0
pip==21.2.4
pyee==8.2.2
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
python-decouple==3.6
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.27.1
requests-html==0.10.0
setuptools==61.2.0
six==1.16.0
soupsieve==2.3.2.post1
tornado==6.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
w3lib==1.22.0
websockets==10.3
wheel==0.37.1
zipp==3.8.0
